+++ agrest-engine/src/main/java/io/agrest/runtime/entity/MapByMerger.java
@@ -5,6 +5,7 @@ import io.agrest.ResourceEntity;
 import io.agrest.RootResourceEntity;
 import io.agrest.ToManyResourceEntity;
 import io.agrest.ToOneResourceEntity;
+import io.agrest.access.MaxPathDepth;
 import io.agrest.meta.AgEntityOverlay;
 import io.agrest.meta.AgSchema;
 import org.apache.cayenne.di.Inject;
@@ -27,7 +28,7 @@ public class MapByMerger implements IMapByMerger {
     }
 
     @Override
-    public <T> void merge(ResourceEntity<T> entity, String mapByPath, Map<Class<?>, AgEntityOverlay<?>> overlays, int maxPathDepth) {
+    public <T> void merge(ResourceEntity<T> entity, String mapByPath, Map<Class<?>, AgEntityOverlay<?>> overlays, MaxPathDepth maxPathDepth) {
         if (mapByPath == null) {
             return;
         }
@@ -46,7 +47,7 @@ public class MapByMerger implements IMapByMerger {
                 ? mapByCompanionEntity((RelatedResourceEntity) entity)
                 : mapByCompanionEntity((RootResourceEntity) entity);
 
-        new ResourceEntityTreeBuilder(mapByCompanionEntity, schema, overlays, maxPathDepth, false)
+        new ResourceEntityTreeBuilder(mapByCompanionEntity, schema, overlays, maxPathDepth.getDepth(), false)
                 .inflatePath(mapByPath);
         entity.mapBy(mapByCompanionEntity);
     }