+++ agrest-engine/src/main/java/io/agrest/runtime/processor/select/stage/SelectCreateResourceEntityStage.java
@@ -60,10 +60,10 @@ public class SelectCreateResourceEntityStage implements Processor<SelectContext<
         AgRequest request = context.getRequest();
 
         sizeMerger.merge(resourceEntity, request.getStart(), request.getLimit());
-        includeMerger.merge(resourceEntity, request.getIncludes(), context.getEntityOverlays(), context.getMaxPathDepth().getDepth());
+        includeMerger.merge(resourceEntity, request.getIncludes(), context.getEntityOverlays(), context.getMaxPathDepth());
         excludeMerger.merge(resourceEntity, request.getExcludes());
-        sortMerger.merge(resourceEntity, request.getSorts());
-        mapByMerger.merge(resourceEntity, request.getMapBy(), context.getEntityOverlays(), context.getMaxPathDepth().getDepth());
+        sortMerger.merge(resourceEntity, request.getSorts(), context.getMaxPathDepth());
+        mapByMerger.merge(resourceEntity, request.getMapBy(), context.getEntityOverlays(), context.getMaxPathDepth());
         expMerger.merge(resourceEntity, request.getExp());
 
         context.setEntity(resourceEntity);