+++ agrest-engine/src/test/java/io/agrest/runtime/entity/IncludeMergerTest.java
@@ -3,6 +3,7 @@ package io.agrest.runtime.entity;
 import io.agrest.RelatedResourceEntity;
 import io.agrest.ResourceEntity;
 import io.agrest.RootResourceEntity;
+import io.agrest.access.MaxPathDepth;
 import io.agrest.annotation.AgAttribute;
 import io.agrest.annotation.AgId;
 import io.agrest.annotation.AgRelationship;
@@ -49,7 +50,7 @@ public class IncludeMergerTest {
 
         AgEntity<X> entity = schema.getEntity(X.class);
         ResourceEntity<X> root = new RootResourceEntity<>(entity);
-        includeMerger.merge(root, asList(), Collections.emptyMap(), 100);
+        includeMerger.merge(root, asList(), Collections.emptyMap(), MaxPathDepth.ofDefault());
 
         assertEquals(entity.getAttributes().size(), root.getBaseProjection().getAttributes().size());
         assertTrue(root.isIdIncluded());
@@ -61,7 +62,7 @@ public class IncludeMergerTest {
 
         AgEntity<X> entity = schema.getEntity(X.class);
         ResourceEntity<X> root = new RootResourceEntity<>(entity);
-        includeMerger.merge(root, asList(new Include("name")), Collections.emptyMap(), 100);
+        includeMerger.merge(root, asList(new Include("name")), Collections.emptyMap(), MaxPathDepth.ofDefault());
 
         assertEquals(1, root.getBaseProjection().getAttributes().size());
         assertFalse(root.isIdIncluded());
@@ -73,7 +74,7 @@ public class IncludeMergerTest {
 
         AgEntity<X> entity = schema.getEntity(X.class);
         ResourceEntity<X> root = new RootResourceEntity<>(entity);
-        includeMerger.merge(root, asList(new Include("name"), new Include("ys")), Collections.emptyMap(), 100);
+        includeMerger.merge(root, asList(new Include("name"), new Include("ys")), Collections.emptyMap(), MaxPathDepth.ofDefault());
 
         assertEquals(1, root.getBaseProjection().getAttributes().size());
         assertFalse(root.isIdIncluded());
@@ -89,7 +90,7 @@ public class IncludeMergerTest {
                 new Include("name"),
                 new Include("ys.name"),
                 new Include("ys.z")
-        ), Collections.emptyMap(), 100);
+        ), Collections.emptyMap(), MaxPathDepth.ofDefault());
 
         assertEquals(1, root.getBaseProjection().getAttributes().size());
         assertFalse(root.isIdIncluded());
@@ -128,7 +129,7 @@ public class IncludeMergerTest {
                 new Include("name"),
                 new Include("ys.name"),
                 new Include("ys.z")
-        ), overlays, 100);
+        ), overlays, MaxPathDepth.ofDefault());
 
         assertEquals(1, root.getBaseProjection().getAttributes().size());
         assertFalse(root.isIdIncluded());