+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #594 Support for escape char in "like" expressions
 * #597 Upgrade Cayenne to 4.2.RC2
 * #598 Max include depth
+* #600 Max include depth for "mapBy", "sort"
 
 ## Release 5.0.M9